refactor: FSD 폴더 구조 1차 반영
