+++ src/shared/ui/index.ts
@@ -1,0 +1,26 @@
+// Button
+export { Button } from "./button"
+
+// Input
+export { Input } from "./input"
+
+// Textarea
+export { Textarea } from "./textarea"
+
+// Card
+export { Card, CardHeader, CardTitle, CardContent } from "./card"
+
+// Table
+export { Table, TableHeader, TableBody, TableRow, TableHead, TableCell } from "./table"
+
+// Header
+export { Header } from "./header"
+
+// Footer
+export { Footer } from "./footer"
+
+// Select
+export { Select, SelectGroup, SelectValue, SelectTrigger, SelectContent, SelectItem } from "./select"
+
+// Dialog
+export { Dialog, DialogTrigger, DialogPortal, DialogOverlay, DialogContent, DialogHeader, DialogTitle } from "./dialog"
